requirements.txt
python-dotenv
opencv-python
google-generativeai